requirements.txt
av
numpy
Pillow
streamlit
streamlit-webrtc
torch
ultralytics
opencv-python-headless
requests
pandas
folium
streamlit-folium
timm
piexif
fpdf2
streamlit-image-comparison
plotly